% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{top_authors}
\alias{top_authors}
\title{Title}
\usage{
top_authors(df, select_category = "all", number = 10, min_date, max_date)
}
\arguments{
\item{df}{}

\item{select_category}{1 or more catagories as a vector}

\item{number}{}
}
\value{

}
\description{
Title
}
